requirements.txt
fastapi
uvicorn[standard]
pillow

numpy==1.26.4

torch==2.1.0
torchvision==0.16.0

ftfy
regex
tqdm

packaging
python-multipart